Refactor: abstract FormLoanWrapper as cypress support
